# contract/openapi.yaml
...
openapi: 3.1.0
info:
  title: VibeCode Store API
  version: 1.0.0
  description: |
    A web store API defined once and implemented twice — NestJS and Spring Boot —
    against one shared PostgreSQL schema, so that a single contract test suite
    can prove the two implementations are interchangeable.

    Conventions that apply to every operation in this file:

    1. Money is an integer count of the currency's minor unit. Field names end
       in `Cents`. Never a float, never a decimal string.
    2. Timestamps are RFC 3339 UTC with millisecond precision and a `Z` suffix:
       `2026-03-14T08:21:05.123Z`.
    3. Public identifiers are lowercase hyphenated UUID strings.
    4. Every error response is `application/problem+json` (RFC 9457). There is
       no second error shape.
    5. Enums are lowercase snake_case strings, never integers.
    6. Unrecognised query parameters are ignored, not rejected.
    7. An absent optional field is omitted from the body, not sent as `null`
       (unless the schema declares it nullable).
    8. Every list operation returns the `Page` envelope, even for one item.

    Three questions this spec used to leave open, now settled:

    - **`sort=price_asc` orders by the product's cheapest active variant**, read
      from a denormalised `min_price_cents` column on `products`. A `MIN()`
      subquery over variants cannot use an index. Both implementations must
      maintain that column whenever a variant price changes or a variant is
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: Browse and search products
      description: >
        Cursor-paginated product list. Only `active` products are ever returned;
        `draft` and `archived` are invisible here regardless of filters.

        `categorySlug` includes the whole subtree — filtering by a parent
        category returns products filed under its descendants too, resolved with
        a recursive CTE. Both implementations must agree on this: selecting a
        top-level category and getting only its directly-filed products is a
        parity failure.

        `q` is a case-insensitive substring match over name and description,
        backed by a trigram index. Not tokenised full-text search: Thai is
        written without spaces between words, so a tsvector of
        `เสื้อเชิ้ตลินินคอปกเปิด` is a single token that `ลินิน` cannot match.
        Trigrams have no notion of a word, which is why they work here. Both
        implementations must match substrings identically, mid-word included.

        `sort=price_asc` and `price_desc` order by the
        product's cheapest active variant, read from the denormalised
        `min_price_cents` column — never a `MIN()` subquery, which cannot use an
        index. Every sort carries `id` as a tiebreaker, without which rows with
        equal sort values duplicate or vanish between pages.

        Unknown query parameters are ignored rather than rejected.
      security: []
      parameters:
        - name: q
          in: query
          required: false
          description: Full-text search over product name and description.
          schema:
            type: string
            minLength: 1
            maxLength: 120
        - name: categorySlug
          in: query
          required: false
          description: Restrict to this category and all of its descendants.
          schema:
            type: string
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        - name: minPriceCents
          in: query
          required: false
          description: Lower bound, inclusive, on the product's cheapest variant.
          schema:
            type: integer
            minimum: 0
        - name: maxPriceCents
          in: query
          required: false
          description: Upper bound, inclusive, on the product's cheapest variant.
          schema:
            type: integer
            minimum: 0
        - name: inStockOnly
          in: query
          required: false
          description: >
            Restrict to products with at least one variant whose available stock
            is above zero. Advisory only — stock can change before checkout.
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          description: Ordering. Defaults to `newest`.
          schema:
            $ref: "#/components/schemas/ProductSort"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: One product, with its variants
      description: >
        Full detail for one active product, including every active variant with
        its available stock and every image.

        `availableStock` is `stock_on_hand - stock_reserved` computed at read
        time. It is display information and may already be stale by the time the
        response is rendered — checkout re-verifies under a row lock. Neither a
        client nor a future reader of this spec should make a sell/no-sell
        decision from it.
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          description: URL-safe product identifier.
          schema:
            type: string
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
      responses:
        "200":
          description: The product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # --- cart ----------------------------------------------------------------
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: The caller's order history
      description: >
        Cursor-paginated, newest first, scoped to the signed-in user. An admin
        calling this sees their own orders, not everyone's — the back-office
        list is `GET /admin/orders`.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: Every product, in any status, with raw stock
      description: >
        The back office's view of the catalogue. Unlike `GET /products` it
        includes `draft` and `archived` products — the admin UI cannot edit what
        it cannot find — and each variant carries `stockOnHand` and
        `stockReserved`, which the inventory screen needs and customers never see.
        Newest first, with the same `Page` envelope and cursor codec as every
        other list.
      parameters:
        - name: status
          in: query
          required: false
          description: Restrict to one publication status.
          schema:
            $ref: "#/components/schemas/ProductStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product with its variants
      description: >
        Creates the product and all of its variants in one transaction — a
        product with no sellable variant is not a useful half-state to allow.
        A duplicate slug or SKU returns 409, never 500. Sets
        `min_price_cents` from the cheapest variant, since `sort=price_asc`
        reads that column.
      responses:
        "201":
          description: The created product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
  /admin/products/{productId}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      description: >
        Partial update; omitted fields are left alone. Moving a product to
        `archived` hides it from the catalogue but never deletes it — order
        history holds `ON DELETE RESTRICT` references to its variants, so the
        rows cannot be removed even deliberately. A duplicate slug returns 409.
      parameters:
        - name: productId
          in: path
          required: true
          description: Product identifier.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: The updated product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/variants/{variantId}/stock:
    patch:
      operationId: adjustVariantStock
      tags: [admin]
      summary: Adjust stock by a signed delta
      description: >
        Takes a signed `delta`, never an absolute quantity. Two admins who each
        read 40 and write an absolute value silently lose one of the changes; two
        who each send `+10` both land. Applied under a row lock.

        Rejected with 409 when the result would leave
        `stock_on_hand < stock_reserved`, because that is the database
        promising goods that are already committed to someone else.

        Returns the variant including `stockOnHand` and `stockReserved`, which
        only admin responses expose.
      parameters:
        - name: variantId
          in: path
          required: true
          description: Variant identifier.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentRequest"
      responses:
        "200":
          description: The variant, with its new stock figures.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVariant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/orders:
    get:
      operationId: listAllOrders
      tags: [admin]
      summary: Every order
      description: >
        Cursor-paginated across all users, newest first, optionally filtered by
        status and user. Uses the same `Page` envelope and the same cursor codec
        as the customer-facing list — one pagination mechanism in the whole API.
      parameters:
        - name: status
          in: query
          required: false
          description: Restrict to one order status.
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: userId
          in: query
          required: false
          description: Restrict to one user's orders.
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/orders/{orderNumber}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Advance an order's status
      description: >
        Forward-only: `paid` → `fulfilled` → `completed`. Every other
        transition returns 409, and the `errors` array names the states that
        *are* reachable from the current one — one entry per state, with
        `field: "status"` and the state itself as `message`. `errors` is
        omitted when the order can go nowhere. That lets the admin UI render its
        buttons from the error response instead of carrying a fourth copy of the
        state machine.

        Cancellation is not a transition here — it is `POST
        /orders/{orderNumber}/cancel`, and it is only valid before payment.
      parameters:
        - $ref: "#/components/parameters/OrderNumber"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: The updated order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            That transition is not allowed. `errors` lists the reachable states.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /auth/login`. Valid for 15 minutes.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: >
        Opaque cursor from a previous response's `nextCursor`. Pass it back
        verbatim; never construct or parse one. A cursor issued by either
        backend must work against the other.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: Page size.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    OrderNumber:
      name: orderNumber
      in: path
      required: true
      description: Human-facing order number, e.g. `ORD-2026-0000001`.
      schema:
        type: string
        pattern: "^ORD-[0-9]{4}-[0-9]{7}$"
...
  schemas:
...
    Page:
      type: object
      description: >
        Cursor-paginated envelope returned by every list operation. `nextCursor`
        is an opaque token — clients must pass it back verbatim and must never
        parse it. It is omitted when there is no further page.
      required: [items, hasMore]
      properties:
        items:
          type: array
          description: The page of results.
          items: {}
        hasMore:
          type: boolean
          description: Whether a further page exists.
        nextCursor:
          type: string
          description: Opaque cursor for the next page. Omitted on the last page.
...
    Problem:
      type: object
      description: >
        RFC 9457 problem details, and the only error shape this API emits.
        `traceId` is always present and always matches the request's log lines.
        `errors` carries field-level detail when the failure is per-field.
        A stack trace never appears in any field.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: Stable URI identifying the problem kind. Clients switch on this.
          examples: ["https://errors.example.com/insufficient-stock"]
        title:
          type: string
          description: Short, human-readable summary of the problem kind.
        status:
          type: integer
          description: HTTP status code, repeated here per RFC 9457.
          minimum: 400
          maximum: 599
        detail:
          type: string
          description: Explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying this occurrence.
        traceId:
          type: string
          description: Correlates this response with the server logs.
        errors:
          type: array
          description: Field-level detail. Present for validation and stock failures.
          items:
            $ref: "#/components/schemas/FieldError"

    # --- auth ------------------------------------------------------------
...
    ProductStatus:
      type: string
      description: >
        Publication state. Only `active` products appear in the catalogue;
        `archived` hides a product without deleting it, so order history stays
        intact.
      enum: [draft, active, archived]
    ProductSort:
      type: string
      description: >
        Ordering for `GET /products`. The price sorts read the denormalised
        `min_price_cents` column, and every option carries `id` as a tiebreaker.
      enum: [newest, price_asc, price_desc, name_asc]
    ProductImage:
      type: object
      description: One product image.
      required: [id, url, position]
      properties:
        id:
          type: string
          format: uuid
          description: Image identifier.
        url:
          type: string
          format: uri
          description: Absolute image URL.
        alt:
          type: string
          description: Alternative text. Omitted when none was supplied.
        position:
          type: integer
          description: Display order, ascending.
    ProductVariant:
      type: object
      description: >
        One buyable variation of a product. `availableStock` is computed at read
        time and is advisory only — checkout re-verifies under a row lock.
      required: [id, sku, name, priceCents, currency, availableStock]
      properties:
        id:
          type: string
          format: uuid
          description: Variant identifier. Also the lock ordering key at checkout.
        sku:
          type: string
          description: Stock-keeping unit. Unique across the catalogue.
        name:
          type: string
          description: What distinguishes this variant, e.g. `Navy / L`.
        priceCents:
          type: integer
          description: Price in minor units. Integer, never a decimal.
          minimum: 0
        currency:
          type: string
          description: ISO 4217 code. Always `THB` in this deployment.
          examples: ["THB"]
        availableStock:
          type: integer
          description: >
            `stock_on_hand - stock_reserved` at read time. Display only — never
            make a sell/no-sell decision from this value.
          minimum: 0
    AdminVariant:
      description: >
        A variant with the raw stock columns exposed. Admin responses only;
        customers see `availableStock` alone.
      allOf:
        - $ref: "#/components/schemas/ProductVariant"
        - type: object
          required: [stockOnHand, stockReserved]
          properties:
            stockOnHand:
              type: integer
              description: Physical units held.
              minimum: 0
            stockReserved:
              type: integer
              description: >
                Units promised to in-flight orders. The schema enforces
                `stockReserved <= stockOnHand`.
              minimum: 0
    ProductSummary:
      type: object
      description: A product as it appears in a list. No variants, no images beyond one.
      required: [id, slug, name, minPriceCents, currency, inStock]
      properties:
        id:
          type: string
          format: uuid
          description: Product identifier.
        slug:
          type: string
          description: URL-safe identifier, used by `GET /products/{slug}`.
        name:
          type: string
          description: Display name.
        minPriceCents:
          type: integer
          description: >
            Cheapest active variant's price, read from the denormalised column
            that `sort=price_asc` orders by.
          minimum: 0
        currency:
          type: string
          description: ISO 4217 code.
          examples: ["THB"]
        inStock:
          type: boolean
          description: Whether any variant currently has available stock. Advisory.
        imageUrl:
          type: string
          format: uri
          description: Primary image. Omitted when the product has no images.
        categorySlug:
          type: string
          description: The category this product is filed under.
    ProductDetail:
      description: >
        Everything needed to render a product page: the summary fields plus all
        active variants and all images.
      allOf:
        - $ref: "#/components/schemas/ProductSummary"
        - type: object
          required: [description, status, variants, images, createdAt]
          properties:
            description:
              type: string
              description: Long-form product copy.
            status:
              $ref: "#/components/schemas/ProductStatus"
            variants:
              type: array
              description: Active variants, cheapest first.
              items:
                $ref: "#/components/schemas/ProductVariant"
            images:
              type: array
              description: Images in display order.
              items:
                $ref: "#/components/schemas/ProductImage"
            createdAt:
              type: string
              format: date-time
              description: When the product was created.
    ProductPage:
      description: A page of products.
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              description: Products in the requested order.
              items:
                $ref: "#/components/schemas/ProductSummary"

    # --- cart ------------------------------------------------------------
...
    OrderStatus:
      type: string
      description: >
        Where an order stands. `pending_payment` on creation; `paid` when the
        webhook lands; then forward-only through `fulfilled` and `completed`.
        `expired` is set by the reservation sweeper when payment never arrived.
      enum:
        - pending_payment
        - paid
        - fulfilled
        - completed
        - cancelled
        - expired
        - payment_failed
...
    Order:
      type: object
      description: >
        A placed order. `cancellable` is computed by the server so that clients
        never carry their own copy of the rule — three implementations of one
        state machine is three places to fix it.
      required:
        [id, orderNumber, status, cancellable, items, shippingAddress,
         subtotalCents, shippingCents, totalCents, currency, payment, placedAt]
      properties:
        id:
          type: string
          format: uuid
          description: Order identifier.
        orderNumber:
          type: string
          description: Human-facing number, used by every order endpoint.
          examples: ["ORD-2026-0000001"]
        status:
          $ref: "#/components/schemas/OrderStatus"
        cancellable:
          type: boolean
          description: >
            Whether `POST /orders/{orderNumber}/cancel` would succeed right now.
            Read this rather than comparing statuses client-side.
        items:
          type: array
          description: The ordered lines, frozen at checkout.
          items:
            $ref: "#/components/schemas/OrderItem"
        shippingAddress:
          $ref: "#/components/schemas/Address"
        subtotalCents:
          type: integer
          description: Sum of the line totals.
          minimum: 0
        shippingCents:
          type: integer
          description: Flat shipping rate applied to this order.
          minimum: 0
        totalCents:
          type: integer
          description: "`subtotalCents + shippingCents`. What the payment charges."
          minimum: 0
        currency:
          type: string
          description: ISO 4217 code.
          examples: ["THB"]
        payment:
          $ref: "#/components/schemas/Payment"
        note:
          type: string
          description: The shopper's message, if one was supplied.
        placedAt:
          type: string
          format: date-time
          description: When the order was created.
        paidAt:
          type: string
          format: date-time
          description: When payment succeeded. Omitted until it does.
        reservationExpiresAt:
          type: string
          format: date-time
          description: >
            When the stock reservation lapses and the sweeper expires this
            order. Omitted once the order is no longer `pending_payment`.
    OrderPage:
      description: A page of orders.
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              description: Orders, newest first.
              items:
                $ref: "#/components/schemas/Order"
...
    CreateProductRequest:
      type: object
      description: >
        A product and its variants, created together. At least one variant is
        required — a product with nothing to sell is not a valid state.
      required: [slug, name, description, categoryId, variants]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          description: Must be unique. A collision returns 409.
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Display name.
        description:
          type: string
          description: Long-form copy.
        categoryId:
          type: string
          format: uuid
          description: The category to file this under.
        status:
          $ref: "#/components/schemas/ProductStatus"
        variants:
          type: array
          description: The variants to create. Sets the product's `min_price_cents`.
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateVariantRequest"
        imageUrls:
          type: array
          description: Image URLs in display order.
          items:
            type: string
            format: uri
    UpdateProductRequest:
      type: object
      description: >
        Partial update. Omitted fields are left unchanged; this endpoint does
        not touch variants, which have their own operations.
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          description: New slug. A collision returns 409.
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: New display name.
        description:
          type: string
          description: New long-form copy.
        categoryId:
          type: string
          format: uuid
          description: Move to a different category.
        status:
          $ref: "#/components/schemas/ProductStatus"
    AdminProduct:
      description: >
        A product as the back office sees it: every field of `ProductDetail`,
        in any status, with each variant's raw stock columns.
      allOf:
        - $ref: "#/components/schemas/ProductDetail"
        - type: object
          required: [categoryId, variants]
          properties:
            categoryId:
              type: string
              format: uuid
              description: The category it is filed under, as the edit form needs it.
            variants:
              type: array
              description: Active variants, cheapest first, with raw stock.
              items:
                $ref: "#/components/schemas/AdminVariant"
    AdminProductPage:
      description: A page of products for the back office.
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              description: Products, newest first.
              items:
                $ref: "#/components/schemas/AdminProduct"
    StockAdjustmentRequest:
      type: object
      description: >
        A signed change to a variant's stock. Absolute values are deliberately
        not accepted: two admins writing absolutes lose one another's work.
      required: [delta, reason]
      properties:
        delta:
          type: integer
          description: >
            Units to add, or a negative number to remove. Rejected with 409 if
            the result would fall below the variant's reserved stock.
          examples: [10]
        reason:
          type: string
          minLength: 1
          maxLength: 200
          description: Why the stock changed. Recorded for audit.
          examples: ["restock"]
    UpdateOrderStatusRequest:
      type: object
      description: The next status for an order. Forward-only.
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
  responses:
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated, but not allowed to do this.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The request conflicts with the current state of the resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: Well-formed but semantically invalid; see `errors`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalError:
      description: Unexpected server error. Never leaks a stack trace.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"